OneDrive - 숙명여자대학교/문서/[25-1산학협력] 일만백만/10k1m-smwu/requirements.txt
streamlit>=1.28.0
anthropic>=0.7.0
google-cloud-vision>=3.4.0
PyMuPDF>=1.23.0
pdfplumber>=0.9.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
numpy>=1.24.0
pathlib>=1.0.1
